revert: keep use-effect-event, native useEffectEvent unusable in forwardRef/memo
Audit of all call sites: useGlobalKeyDown and useClickOutsideEvent are
public hooks that execute in the caller's fiber (all internal callers —
DialogCard, Menu, Breadcrumbs — are forwardRef components, and external
callers can be anything), and Tooltip is a forwardRef component. React
19.2's native useEffectEvent never updates past first-render values in
forwardRef/memo fibers (facebook/react#34818), so no site can use the
native hook until the fix ships in a stable release.

Restore the use-effect-event dependency at all three sites with TODO
comments, drop the internal shim, and add a regression test that pins
the fresh-values contract for forwardRef/memo callers.
