+++ packages/ui/src/core/primitives/tooltip/tooltip.tsx
@@ -22,10 +22,15 @@ import {
   useState,
 } from 'react'
 import {styled} from 'styled-components'
+// TODO: switch to `useEffectEvent` from `react` once
+// https://github.com/facebook/react/issues/34818 is fixed in the lowest React
+// version we support: on React 19.2 the native hook never sees values past
+// the first render when the calling component is wrapped in `forwardRef` or
+// `memo`, and `Tooltip` is a `forwardRef` component.
+import {useEffectEvent} from 'use-effect-event'
 
 import type {ThemeColorSchemeKey} from '../../../theme/system/color/_system'
 import {useDelayedState} from '../../hooks/useDelayedState'
-import {useEffectEvent} from '../../hooks/useEffectEvent'
 import {usePrefersReducedMotion} from '../../hooks/usePrefersReducedMotion'
 import {origin} from '../../middleware/origin'
 import {_getArrayProp} from '../../styles/helpers'
